+++ CHANGELOG.md
@@ -6,9 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### :rocket: Added
-### :pencil: Changed
-### :bug: Bug Fix
+### :nail: Polish
+- Catch YAML parsing error and display them in the console
 
 ## [1.0.1]
 ### :bug: Bug Fix